requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
onnxruntime==1.18.0
pydantic==2.8.0
